+++ bots/__init__.py
@@ -1,8 +1,0 @@
-"""
-Bots package containing AI bot implementations.
-"""
-
-from .ai_bot import AiBot
-from .gemini_bot import GeminiBot
-
-__all__ = ['AiBot', 'GeminiBot']